Revert fix for URL encoding
