+++ Package.resolved
@@ -5,18 +5,17 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/pointfreeco/swift-snapshot-testing",
       "state" : {
-          "revision": "506b6052384d8e97a4bb16fe8680325351c23c64",
-          "version": "1.14.0"
-        }
-      },
-      {
-        "package": "swift-syntax",
-        "repositoryURL": "https://github.com/apple/swift-syntax.git",
-        "state": {
-          "branch": null,
-          "revision": "74203046135342e4a4a627476dd6caf8b28fe11b",
-          "version": "509.0.0"
-        }
+        "revision" : "506b6052384d8e97a4bb16fe8680325351c23c64",
+        "version" : "1.14.0"
+      }
+    },
+    {
+      "identity" : "swift-syntax",
+      "kind" : "remoteSourceControl",
+      "location" : "https://github.com/apple/swift-syntax.git",
+      "state" : {
+        "revision" : "74203046135342e4a4a627476dd6caf8b28fe11b",
+        "version" : "509.0.0"
       }
     }
   ],
