+++ Sources/PostgREST/PostgrestBuilder.swift
@@ -92,6 +92,9 @@ public class PostgrestBuilder {
       }
     }
 
+    if headers["Accept"] == nil {
+      headers["Accept"] = "application/json"
+    }
     headers["Content-Type"] = "application/json"
 
     if let schema = configuration.schema {
@@ -124,7 +127,9 @@ public class PostgrestBuilder {
     }
 
     if !queryParams.isEmpty {
-      components.queryItems = queryParams.map(URLQueryItem.init)
+      let percentEncodedQuery =
+        (components.percentEncodedQuery.map { $0 + "&" } ?? "") + self.query(queryParams)
+      components.percentEncodedQuery = percentEncodedQuery
     }
 
     guard let url = components.url else {
@@ -145,4 +150,47 @@ public class PostgrestBuilder {
 
     return urlRequest
   }
+
+  private func escape(_ string: String) -> String {
+    string.addingPercentEncoding(withAllowedCharacters: .postgrestURLQueryAllowed) ?? string
+  }
+
+  private func query(_ parameters: [(String, String?)]) -> String {
+    parameters.compactMap { key, value in
+      if let value {
+        return (key, value)
+      }
+      return nil
+    }
+    .map { key, value in
+      let escapedKey = escape(key)
+      let escapedValue = escape(value)
+      return "\(escapedKey)=\(escapedValue)"
+    }
+    .joined(separator: "&")
+  }
+}
+
+extension CharacterSet {
+  /// Creates a CharacterSet from RFC 3986 allowed characters.
+  ///
+  /// RFC 3986 states that the following characters are "reserved" characters.
+  ///
+  /// - General Delimiters: ":", "#", "[", "]", "@", "?", "/"
+  /// - Sub-Delimiters: "!", "$", "&", "'", "(", ")", "*", "+", ",", ";", "="
+  ///
+  /// In RFC 3986 - Section 3.4, it states that the "?" and "/" characters should not be escaped to
+  /// allow
+  /// query strings to include a URL. Therefore, all "reserved" characters with the exception of "?"
+  /// and "/"
+  /// should be percent-escaped in the query string.
+  static let postgrestURLQueryAllowed: CharacterSet = {
+    let generalDelimitersToEncode =
+      ":#[]@"  // does not include "?" or "/" due to RFC 3986 - Section 3.4
+    let subDelimitersToEncode = "!$&'()*+,;="
+    let encodableDelimiters =
+      CharacterSet(charactersIn: "\(generalDelimitersToEncode)\(subDelimitersToEncode)")
+
+    return CharacterSet.urlQueryAllowed.subtracting(encodableDelimiters)
+  }()
 }
